Added auth protection to routes

+++ routes/web.php
@@ -17,14 +17,14 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', [ListingController::class, 'index']);
 Route::post('listings/', [ListingController::class, 'store']);
-Route::get('listings/create', [ListingController::class, 'create']);
+Route::get('listings/create', [ListingController::class, 'create'])->middleware('auth');
 Route::get('listings/{listing}', [ListingController::class, 'show']);
-Route::get('listings/{listing}/edit', [ListingController::class, 'edit']);
-Route::put('listings/{listing}', [ListingController::class, 'update']);
-Route::delete('listings/{listing}', [ListingController::class, 'destroy']);
+Route::get('listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
+Route::put('listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
+Route::delete('listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
-Route::get('register', [UserController::class, 'register']);
-Route::post('users', [UserController::class, 'store']);
-Route::post('logout', [UserController::class, 'logout']);
-Route::get('login', [UserController::class, 'login']);
-Route::post('users/authenticate', [UserController::class, 'authenticate']);
+Route::get('register', [UserController::class, 'register'])->middleware('guest');
+Route::post('users', [UserController::class, 'store'])->middleware('guest');
+Route::post('logout', [UserController::class, 'logout'])->middleware('auth');
+Route::get('login', [UserController::class, 'login'])->middleware('guest')->name('login');
+Route::post('users/authenticate', [UserController::class, 'authenticate'])->middleware('guest');
